Resolves the OGC-1145 conflict in buildSampleEditItems on top of
resolveTestsFromCodeableConcept. Drops the uncalled
createResultUpdateDataSetFromReport transform: a DiagnosticReport is a
projection of Analysis + Results, so writes flow through ServiceRequest
and Observation. Delete now validates the id (400 on a non-UUID),
injects IStatusService, and returns the shared delete outcome;
safeMessage lives in FhirProviderUtils. Facade tests cover
cancel-on-delete, 404 and 400.

+++ src/main/java/org/openelisglobal/common/rest/provider/bean/homedashboard/AverageTimeDisplayBean.java
@@ -2,33 +2,33 @@ package org.openelisglobal.common.rest.provider.bean.homedashboard;
 
 public class AverageTimeDisplayBean {
 
-    Double receptionToResult = 0.0;
+    Long receptionToResult = 0L;
 
-    Double resultToValidation = 0.0;
+    Long resultToValidation = 0L;
 
-    Double receptionToValidation = 0.0;
+    Long receptionToValidation = 0L;
 
-    public Double getReceptionToResult() {
+    public Long getReceptionToResult() {
         return receptionToResult;
     }
 
-    public void setReceptionToResult(Double receptionToResult) {
+    public void setReceptionToResult(Long receptionToResult) {
         this.receptionToResult = receptionToResult;
     }
 
-    public Double getResultToValidation() {
+    public Long getResultToValidation() {
         return resultToValidation;
     }
 
-    public void setResultToValidation(Double resultToValidation) {
+    public void setResultToValidation(Long resultToValidation) {
         this.resultToValidation = resultToValidation;
     }
 
-    public Double getReceptionToValidation() {
+    public Long getReceptionToValidation() {
         return receptionToValidation;
     }
 
-    public void setReceptionToValidation(Double receptionToValidation) {
+    public void setReceptionToValidation(Long receptionToValidation) {
         this.receptionToValidation = receptionToValidation;
     }
 }
